requirements.txt
###### Requirements without Version Specifiers #####
nextcord
pytest
python-dotenv
